separate out the zarr reading code as a separate reader
