+++ virtualizarr/backend.py
@@ -2,7 +2,6 @@ import os
 import warnings
 from collections.abc import Iterable, Mapping, MutableMapping
 from io import BufferedIOBase
-from pathlib import Path
 from typing import (
     Any,
     Hashable,
@@ -20,10 +19,6 @@ import virtualizarr.kerchunk as kerchunk
 from virtualizarr.kerchunk import FileType, KerchunkStoreRefs
 from virtualizarr.manifests import ChunkManifest, ManifestArray
 from virtualizarr.utils import _fsspec_openfile_from_filepath
-from virtualizarr.zarr import (
-    attrs_from_zarr_group_json,
-    metadata_from_zarr_json,
-)
 
 XArrayOpenT = str | os.PathLike[Any] | BufferedIOBase | AbstractDataStore
 
@@ -128,6 +123,8 @@ def open_virtual_dataset(
 
     if filetype == FileType.zarr_v3:
         # TODO is there a neat way of auto-detecting this?
+        from virtualizarr.readers.zarr import open_virtual_dataset_from_v3_store
+
         return open_virtual_dataset_from_v3_store(
             storepath=filepath, drop_variables=drop_variables, indexes=indexes
         )
@@ -227,58 +224,6 @@ def open_virtual_dataset(
         # TODO we should probably also use vds.set_close() to tell xarray how to close the file we opened
 
         return vds
-
-
-def open_virtual_dataset_from_v3_store(
-    storepath: str,
-    drop_variables: list[str],
-    indexes: Mapping[str, Index] | None,
-) -> xr.Dataset:
-    """
-    Read a Zarr v3 store and return an xarray Dataset containing virtualized arrays.
-    """
-    _storepath = Path(storepath)
-
-    ds_attrs = attrs_from_zarr_group_json(_storepath / "zarr.json")
-    coord_names = ds_attrs.pop("coordinates", [])
-
-    # TODO recursive glob to create a datatree
-    # Note: this .is_file() check should not be necessary according to the pathlib docs, but tests fail on github CI without it
-    # see https://github.com/TomNicholas/VirtualiZarr/pull/45#discussion_r1547833166
-    all_paths = _storepath.glob("*/")
-    directory_paths = [p for p in all_paths if not p.is_file()]
-
-    vars = {}
-    for array_dir in directory_paths:
-        var_name = array_dir.name
-        if var_name in drop_variables:
-            break
-
-        zarray, dim_names, attrs = metadata_from_zarr_json(array_dir / "zarr.json")
-        manifest = ChunkManifest.from_zarr_json(str(array_dir / "manifest.json"))
-
-        marr = ManifestArray(chunkmanifest=manifest, zarray=zarray)
-        var = xr.Variable(data=marr, dims=dim_names, attrs=attrs)
-        vars[var_name] = var
-
-    if indexes is None:
-        raise NotImplementedError()
-    elif indexes != {}:
-        # TODO allow manual specification of index objects
-        raise NotImplementedError()
-    else:
-        indexes = dict(**indexes)  # for type hinting: to allow mutation
-
-    data_vars, coords = separate_coords(vars, indexes, coord_names)
-
-    ds = xr.Dataset(
-        data_vars,
-        coords=coords,
-        # indexes={},  # TODO should be added in a later version of xarray
-        attrs=ds_attrs,
-    )
-
-    return ds
 
 
 def virtual_vars_from_kerchunk_refs(
