+++ virtualizarr/zarr.py
@@ -1,14 +1,9 @@
 import dataclasses
-import json
-from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, NewType, cast
 
 import numcodecs
 import numpy as np
 import ujson  # type: ignore
-import xarray as xr
-
-from virtualizarr.vendor.zarr.utils import json_dumps
 
 if TYPE_CHECKING:
     pass
@@ -211,179 +206,6 @@ def ceildiv(a: int, b: int) -> int:
     See https://stackoverflow.com/questions/14822184/is-there-a-ceiling-equivalent-of-operator-in-python
     """
     return -(a // -b)
-
-
-def dataset_to_zarr(ds: xr.Dataset, storepath: str) -> None:
-    """
-    Write an xarray dataset whose variables wrap ManifestArrays to a v3 Zarr store, writing chunk references into manifest.json files.
-
-    Currently requires all variables to be backed by ManifestArray objects.
-
-    Not very useful until some implementation of a Zarr reader can actually read these manifest.json files.
-    See https://github.com/zarr-developers/zarr-specs/issues/287
-
-    Parameters
-    ----------
-    ds: xr.Dataset
-    storepath: str
-    """
-
-    from virtualizarr.manifests import ManifestArray
-
-    _storepath = Path(storepath)
-    Path.mkdir(_storepath, exist_ok=False)
-
-    # should techically loop over groups in a tree but a dataset corresponds to only one group
-    group_metadata = {"zarr_format": 3, "node_type": "group", "attributes": ds.attrs}
-    with open(_storepath / "zarr.json", "wb") as group_metadata_file:
-        group_metadata_file.write(json_dumps(group_metadata))
-
-    for name, var in ds.variables.items():
-        array_dir = _storepath / str(name)
-        marr = var.data
-
-        # TODO move this check outside the writing loop so we don't write an incomplete store on failure?
-        # TODO at some point this should be generalized to also write in-memory arrays as normal zarr chunks, see GH isse #62.
-        if not isinstance(marr, ManifestArray):
-            raise TypeError(
-                "Only xarray objects wrapping ManifestArrays can be written to zarr using this method, "
-                f"but variable {name} wraps an array of type {type(marr)}"
-            )
-
-        Path.mkdir(array_dir, exist_ok=False)
-
-        # write the chunk references into a manifest.json file
-        # and the array metadata into a zarr.json file
-        to_zarr_json(var, array_dir)
-
-
-def to_zarr_json(var: xr.Variable, array_dir: Path) -> None:
-    """
-    Write out both the zarr.json and manifest.json file into the given zarr array directory.
-
-    Follows the Zarr v3 manifest storage transformer ZEP (see https://github.com/zarr-developers/zarr-specs/issues/287).
-
-    Parameters
-    ----------
-    var : xr.Variable
-        Must be wrapping a ManifestArray
-    dirpath : str
-        Zarr store array directory into which to write files.
-    """
-
-    marr = var.data
-
-    marr.manifest.to_zarr_json(array_dir / "manifest.json")
-
-    metadata = zarr_v3_array_metadata(
-        marr.zarray, [str(x) for x in var.dims], var.attrs
-    )
-    with open(array_dir / "zarr.json", "wb") as metadata_file:
-        metadata_file.write(json_dumps(metadata))
-
-
-def zarr_v3_array_metadata(zarray: ZArray, dim_names: list[str], attrs: dict) -> dict:
-    """Construct a v3-compliant metadata dict from v2 zarray + information stored on the xarray variable."""
-    # TODO it would be nice if we could use the zarr-python metadata.ArrayMetadata classes to do this conversion for us
-
-    metadata = zarray.dict()
-
-    # adjust to match v3 spec
-    metadata["zarr_format"] = 3
-    metadata["node_type"] = "array"
-    metadata["data_type"] = str(np.dtype(metadata.pop("dtype")))
-    metadata["chunk_grid"] = {
-        "name": "regular",
-        "configuration": {"chunk_shape": metadata.pop("chunks")},
-    }
-    metadata["chunk_key_encoding"] = {
-        "name": "default",
-        "configuration": {"separator": "/"},
-    }
-    metadata["codecs"] = zarray._v3_codec_pipeline()
-    metadata.pop("filters")
-    metadata.pop("compressor")
-    metadata.pop("order")
-
-    # indicate that we're using the manifest storage transformer ZEP
-    metadata["storage_transformers"] = [
-        {
-            "name": "chunk-manifest-json",
-            "configuration": {"manifest": "./manifest.json"},
-        }
-    ]
-
-    # add information from xarray object
-    metadata["dimension_names"] = dim_names
-    metadata["attributes"] = attrs
-
-    return metadata
-
-
-def attrs_from_zarr_group_json(filepath: Path) -> dict:
-    with open(filepath) as metadata_file:
-        attrs = json.load(metadata_file)
-    return attrs["attributes"]
-
-
-def metadata_from_zarr_json(filepath: Path) -> tuple[ZArray, list[str], dict]:
-    with open(filepath) as metadata_file:
-        metadata = json.load(metadata_file)
-
-    if {
-        "name": "chunk-manifest-json",
-        "configuration": {
-            "manifest": "./manifest.json",
-        },
-    } not in metadata.get("storage_transformers", []):
-        raise ValueError(
-            "Can only read byte ranges from Zarr v3 stores which implement the manifest storage transformer ZEP."
-        )
-
-    attrs = metadata.pop("attributes")
-    dim_names = metadata.pop("dimension_names")
-
-    chunk_shape = tuple(metadata["chunk_grid"]["configuration"]["chunk_shape"])
-    shape = tuple(metadata["shape"])
-    zarr_format = metadata["zarr_format"]
-
-    if metadata["fill_value"] is None:
-        raise ValueError(
-            "fill_value must be specified https://zarr-specs.readthedocs.io/en/latest/v3/core/v3.0.html#fill-value"
-        )
-    else:
-        fill_value = metadata["fill_value"]
-
-    all_codecs = [
-        codec
-        for codec in metadata["codecs"]
-        if codec["name"] not in ("transpose", "bytes")
-    ]
-    compressor, *filters = [
-        _configurable_to_num_codec_config(_filter) for _filter in all_codecs
-    ]
-    zarray = ZArray(
-        chunks=chunk_shape,
-        compressor=compressor,
-        dtype=np.dtype(metadata["data_type"]),
-        fill_value=fill_value,
-        filters=filters or None,
-        order="C",
-        shape=shape,
-        zarr_format=zarr_format,
-    )
-
-    return zarray, dim_names, attrs
-
-
-def _configurable_to_num_codec_config(configurable: dict) -> dict:
-    """
-    Convert a zarr v3 configurable into a numcodecs codec.
-    """
-    configurable_copy = configurable.copy()
-    codec_id = configurable_copy.pop("name")
-    configuration = configurable_copy.pop("configuration")
-    return numcodecs.get_codec({"id": codec_id, **configuration}).get_config()
 
 
 def _num_codec_config_to_configurable(num_codec: dict) -> dict:
